Change default PCA kernel size
Previously the default spherical kernel targeted a fixed aspect ratio. With this change, the estimated rank from a prior iteration is used to modulate the size of the PCA kernel at each location. The premise here is that, excluding signal components, the noise-only portion of the eigenspectrum should be approximately square.