New option -minvoxels
The minimum Casorati matrix aspect ratio is used by default to define the size of an increasing-radius spherically-shaped kernel in order for some minimum number of voxels to be present. This change adds a new command-line option -minvoxels to instead control this number directly.
Further, options -radius_mm and -radius_ratio have been renamed to -radius and -aspect_ratio respectively.