Refactoring of multi-threaded operation
Previously, all threads possessed access to the output denoised image, and would mutex lock across all threads before accumulating results there. This may however be resulting in considerable delays in the acquisition and release of these locks due to the need to synchronize the memory view of a potentially very large output image across all threads.
In this change, processing is done via a Thread::run_queue() call, and the morker threads pass just the outcomes of the denoising of each patch individually down a multi-thread-safe queue, for a single receiver thread to aggregate the denoised image data.